refactor feign config
