+++ projects/prison-case-notes-to-probation/src/main/kotlin/uk/gov/justice/digital/hmpps/config/CaseNotesFeignConfig.kt
@@ -1,0 +1,15 @@
+package uk.gov.justice.digital.hmpps.config
+
+import org.springframework.cloud.openfeign.EnableFeignClients
+import org.springframework.context.annotation.Configuration
+import org.springframework.security.oauth2.client.OAuth2AuthorizedClientManager
+import uk.gov.justice.digital.hmpps.feign.FeignConfig
+import uk.gov.justice.digital.hmpps.integrations.prison.PrisonCaseNotesClient
+
+@Configuration
+@EnableFeignClients(clients = [PrisonCaseNotesClient::class])
+class CaseNotesFeignConfig(
+    authorizedClientManager: OAuth2AuthorizedClientManager
+) : FeignConfig(authorizedClientManager) {
+    override fun registrationId() = "prison-case-notes"
+}
