+++ projects/workforce-allocations-to-delius/src/main/kotlin/uk/gov/justice/digital/hmpps/config/WorkForceFeignConfig.kt
@@ -1,0 +1,15 @@
+package uk.gov.justice.digital.hmpps.config
+
+import org.springframework.cloud.openfeign.EnableFeignClients
+import org.springframework.context.annotation.Configuration
+import org.springframework.security.oauth2.client.OAuth2AuthorizedClientManager
+import uk.gov.justice.digital.hmpps.feign.FeignConfig
+import uk.gov.justice.digital.hmpps.integrations.workforceallocations.WorkforceAllocationsClient
+
+@Configuration
+@EnableFeignClients(clients = [WorkforceAllocationsClient::class])
+class WorkForceFeignConfig(
+    authorizedClientManager: OAuth2AuthorizedClientManager
+) : FeignConfig(authorizedClientManager) {
+    override fun registrationId() = "workforce-allocations-to-delius"
+}
